fix: WSL 解析测试移入 Windows 构建标签文件,修复 Linux CI 未定义符号
parseWSLDistros 定义于 //go:build windows 的 localshell_windows.go,
但其测试位于无标签的 localservice_test.go,Linux CI 编译时符号未定义。
测试移至 localshell_windows_test.go(同标签),与实现保持同一构建域。
